Tailwind added and all the functionality move to main componenets
