+++ package.json
@@ -14,7 +14,9 @@
     "start": "npm run compile && electron .",
     "compile": "tsc",
     "check": "svelte-check --tsconfig ./tsconfig.json",
-    "bin": "npm run out:win && npm run out:mac && npm run out:linux",
+    "test": "jest src",
+    "test:watch": "yarn  test -- --watch",
+    "bin": "yarn out:win && npm run out:mac && npm run out:linux",
     "bin:win": "rollup -c && tsc && electron-builder build --win --publish never",
     "bin:mac": "rollup -c && tsc && electron-builder build --mac --publish never",
     "bin:linux": "rollup -c && tsc && electron-builder build --linux --publish never"
@@ -23,10 +25,15 @@
     "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-node-resolve": "^13.0.6",
     "@rollup/plugin-typescript": "^8.3.0",
+    "@testing-library/svelte": "^3.0.3",
     "@tsconfig/svelte": "^2.0.0",
+    "autoprefixer": "^10.4.0",
+    "babel-jest": "^27.3.1",
     "electron": "^15.2.0",
     "electron-builder": "^22.13.1",
+    "jest": "^27.3.1",
     "nodemon": "^2.0.14",
+    "postcss": "^8.3.11",
     "rollup": "^2.58.0",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-livereload": "^2.0.5",
@@ -34,7 +41,9 @@
     "rollup-plugin-terser": "^7.0.2",
     "svelte": "^3.0.0",
     "svelte-check": "^2.0.0",
-    "svelte-preprocess": "^4.0.0",
+    "svelte-jester": "^2.1.5",
+    "svelte-preprocess": "^4.9.8",
+    "tailwindcss": "^2.2.19",
     "tslib": "^2.0.0",
     "typescript": "^4.4.4"
   },
@@ -42,5 +51,19 @@
     "@iota/iota.js": "^1.8.1",
     "electron-reload": "^2.0.0-alpha.1",
     "sirv-cli": "^1.0.0"
+  },
+  "jest": {
+    "verbose": true,
+    "transform": {
+      "^.+\\.js$": "babel-jest",
+      "^.+\\.svelte$": "svelte-jester"
+    },
+    "moduleFileExtensions": [
+      "js",
+      "svelte"
+    ],
+    "setupFilesAfterEnv": [
+      "@testing-library/jest-dom/extend-expect"
+    ]
   }
 }
